Update audiorecorder.py
In normal state (100 <= max_value <= 300), then it records the current input waveform.
E.g. in this program, if the current state is normal, then it returns 0.
If the max_value of the current input exceeds 300, moving average method is activated, if the current rms_value is less than 300, it returns 0, e.g. normal state.
If not, it represent third state and return 1.
If program detects silence state, it increases the num_silent value.
If the num_silent value is equal to 200, it stops input and terminate. Also saves wav file.
Terminate condition (num_silent) also can be adjustable.
The last length parameter represents the length of current input buffer. You can adjust its value you want.